+++ test/groundwater.jl
@@ -46,6 +46,7 @@ function homogenous_aquifer(nrow, ncol)
         fill(0.1, ncell), # specific storage
         fill(1.0, ncell),  # storativity
         fill(0.0, connectivity.nconnection),  # conductance
+        fill(0.0, ncell), # total volume that can be released
     )
     unconf_aqf = Wflow.UnconfinedAquifer(
         [0.0, 7.5, 20.0],  # head
@@ -55,6 +56,7 @@ function homogenous_aquifer(nrow, ncol)
         fill(100.0, ncell),  # area
         fill(0.15, ncell),  # specific yield
         fill(0.0, connectivity.nconnection),  # conductance
+        fill(0.0, ncell), # total volume that can be released
         fill(3.0, ncell), # conductance reduction factor
     )
     return (connectivity, conf_aqf, unconf_aqf)
@@ -411,6 +413,7 @@ end
             fill(cellsize * cellsize, ncell),
             fill(specific_yield, ncell),
             fill(0.0, connectivity.nconnection),
+            fill(0.0, ncell),
             fill(gwf_f, ncell),
         )
         # constant head on left boundary, 0 at 0
@@ -435,7 +438,14 @@ end
         end
 
         head_analytical = [
-            transient_aquifer_1d(x, time, conductivity, specific_yield, aquifer_length, beta) for x in xc
+            transient_aquifer_1d(
+                x,
+                time,
+                conductivity,
+                specific_yield,
+                aquifer_length,
+                beta,
+            ) for x in xc
         ]
         difference = gwf.aquifer.head .- head_analytical
         # @test all(difference .< ?)  #TODO
@@ -471,6 +481,7 @@ end
             fill(cellsize * cellsize, ncell),
             fill(specific_yield, ncell),
             fill(0.0, connectivity.nconnection),
+            fill(0.0, ncell),
             fill(gwf_f, ncell),
         )
         # constant head on left boundary, 0 at 0
@@ -495,7 +506,14 @@ end
         end
 
         head_analytical = [
-            transient_aquifer_1d(x, time, conductivity, specific_yield, aquifer_length, beta) for x in xc
+            transient_aquifer_1d(
+                x,
+                time,
+                conductivity,
+                specific_yield,
+                aquifer_length,
+                beta,
+            ) for x in xc
         ]
         difference = gwf.aquifer.head .- head_analytical
         # @test all(difference .< ?)  #TODO
@@ -535,6 +553,7 @@ end
             fill(specific_storage, ncell),
             fill(storativity, ncell),
             fill(0.0, connectivity.nconnection), # conductance, to be set
+            fill(0.0, ncell), # total volume that can be released, to be set
         )
 
         cell_index = reshape(collect(range(1, ncell, step = 1)), shape)