+++ test/run.jl
@@ -3,9 +3,6 @@ using Wflow
 using Dates
 using NCDatasets
 
-tomlpath = joinpath(@__DIR__, "sbm_simple.toml")
-Wflow.run(tomlpath; silent = true)
-
 # test whether restarted runs get the same results as continuous ones, i.e. state is captured
 tomlpath = joinpath(@__DIR__, "sbm_config.toml")
 config = Wflow.Config(tomlpath)