+++ test/vertical_process.jl
@@ -1,3 +1,4 @@
+using Dates
 @testset "vertical processes" begin
     @test all(
         isapprox.(
@@ -11,28 +12,22 @@
             (4.343, 3.87, 0.387, 0.0, 3.8, 2.043),
         ),
     )
-    @test all(
-        isapprox.(
-            Wflow.acttransp_unsat_sbm(
-                300.0,
-                55.0,
-                400.0,
-                3.6,
-                1.85,
-                10.5,
-                300.0,
-                0.6,
-                0.1,
-                10.0,
-                false,
-            ),
-            (55.0, 1.85, 3.6),
-        ),
-    )
+    @test Wflow.head_brooks_corey(0.25, 0.6, 0.15, 10.5, -10.0) ≈ -90.6299820833844
+    @test Wflow.feddes_h3(-300.0, -600.0, 3.5, Second(86400)) ≈ -412.5
+    @test Wflow.feddes_h3(-300.0, -600.0, 0.5, Second(86400)) == -600.0
+    @test Wflow.feddes_h3(-300.0, -600.0, 6.0, Second(86400)) == -300.0
+    @test Wflow.rwu_reduction_feddes(0.0, -10.0, -100.0, -300.0, -15000.0, 0.0) == 0.0
+    @test Wflow.rwu_reduction_feddes(0.0, -10.0, -100.0, -300.0, -15000.0, 1.0) == 1.0
+    @test Wflow.rwu_reduction_feddes(-90.0, -10.0, -100.0, -412.5, -15000.0, 0.0) ≈
+          0.8888888888888888
+    @test Wflow.rwu_reduction_feddes(-350.0, -10.0, -100.0, -412.5, -15000.0, 0.0) == 1.0
+    @test Wflow.rwu_reduction_feddes(-12000.0, -10.0, -100.0, -412.5, -15000.0, 0.0) ≈
+          0.20565552699228792
+    @test Wflow.rwu_reduction_feddes(-16000.0, -10.0, -100.0, -412.5, -15000.0, 0.0) == 0.0
     @test all(
         isapprox.(
             Wflow.infiltration(27.5, 0.2, 0.038, 8.9, 50.0, 5.0, 23.5, false, false),
-            (23.5, 19.14814814814815, 4.351851851851852, 22.0, 5.5, 0.5),
+            (23.5, 19.14814814814815, 4.351851851851852, 22.0, 5.5, 0.5, 1.0),
         ),
     )
     @test all(